+++ resolwe/test/__init__.py
@@ -16,12 +16,16 @@ from resolwe.test.testcases import TestCase, TransactionTestCase
 from resolwe.test.testcases.api import ResolweAPITestCase
 from resolwe.test.testcases.elastic import ElasticSearchTestCase, TransactionElasticSearchTestCase
 from resolwe.test.testcases.process import ProcessTestCase, TransactionProcessTestCase
-from resolwe.test.utils import check_docker, check_installed, with_docker_executor, with_null_executor
+from resolwe.test.utils import (
+    check_docker, check_installed, with_custom_executor, with_docker_executor, with_null_executor,
+    with_resolwe_host,
+)
 
 __all__ = (
     'TestCase', 'TransactionTestCase',
     'ResolweAPITestCase',
     'ElasticSearchTestCase', 'TransactionElasticSearchTestCase',
     'ProcessTestCase', 'TransactionProcessTestCase',
-    'check_docker', 'check_installed', 'with_docker_executor',
+    'check_docker', 'check_installed', 'with_custom_executor', 'with_docker_executor',
+    'with_null_executor', 'with_resolwe_host',
 )