+++ resolwe/test/utils.py
@@ -7,6 +7,7 @@ Resolwe Test Helpers and Decorators
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import functools
 import os
 import shlex
 import shutil
@@ -14,6 +15,7 @@ import subprocess
 import unittest
 
 import six
+import wrapt
 
 from django.conf import settings
 from django.test import override_settings
@@ -24,7 +26,10 @@ if six.PY2:
     # Monkey-patch shutil package with which function (available in Python 3.3+)
     import shutilwhich  # pylint: disable=import-error,unused-import
 
-__all__ = ('check_installed', 'check_docker', 'with_docker_executor')
+__all__ = (
+    'check_installed', 'check_docker', 'with_custom_executor', 'with_docker_executor',
+    'with_null_executor', 'with_resolwe_host',
+)
 
 
 def check_installed(command):
@@ -66,16 +71,21 @@ def check_docker():
     return available, reason
 
 
-def with_docker_executor(method):
-    """Decorate unit test to run processes with Docker executor."""
+def with_custom_executor(wrapped=None, **custom_executor_settings):
+    """Decorate unit test to run processes with a custom executor.
+
+    :param dict custom_executor_settings: custom ``FLOW_EXECUTOR``
+        settings with which you wish to override the current settings
+
+    """
+    if wrapped is None:
+        return functools.partial(with_custom_executor, **custom_executor_settings)
+
     # pylint: disable=missing-docstring
-    @unittest.skipUnless(*check_docker())
-    def wrapper(*args, **kwargs):
+    @wrapt.decorator
+    def wrapper(wrapped_method, instance, args, kwargs):
         executor_settings = settings.FLOW_EXECUTOR.copy()
-        executor_settings.update({
-            'NAME': 'resolwe.flow.executors.docker',
-            'CONTAINER_IMAGE': 'resolwe/base',
-        })
+        executor_settings.update(custom_executor_settings)
 
         try:
             with override_settings(FLOW_EXECUTOR=executor_settings):
@@ -83,32 +93,65 @@ def with_docker_executor(method):
                 manager.discover_engines()
 
                 # Run the actual unit test method.
-                method(*args, **kwargs)
+                return wrapped_method(*args, **kwargs)
         finally:
             # Re-run engine discovery as the settings have changed.
             manager.discover_engines()
 
-    return wrapper
+    return wrapper(wrapped)  # pylint: disable=no-value-for-parameter
 
 
-def with_null_executor(method):
-    """Decorate unit test to run processes with Null executor."""
+def with_docker_executor(wrapped=None, image='resolwe/base'):
+    """Decorate unit test to run processes with the Docker executor.
+
+    :param str image: custom Docker image to use when running the Docker
+        executor
+
+    """
+    if wrapped is None:
+        return functools.partial(with_docker_executor, image=image)
+
     # pylint: disable=missing-docstring
-    def wrapper(*args, **kwargs):
-        executor_settings = settings.FLOW_EXECUTOR.copy()
-        executor_settings.update({
-            'NAME': 'resolwe.flow.executors.null',
-        })
+    @wrapt.decorator
+    def wrapper(wrapped_method, instance, args, kwargs):
+        return unittest.skipUnless(*check_docker())(
+            with_custom_executor(
+                NAME='resolwe.flow.executors.docker',
+                CONTAINER_IMAGE=image,
+            )(wrapped_method)
+        )(*args, **kwargs)
 
-        try:
-            with override_settings(FLOW_EXECUTOR=executor_settings):
-                # Re-run engine discovery as the settings have changed.
-                manager.discover_engines()
+    return wrapper(wrapped)  # pylint: disable=no-value-for-parameter
 
-                # Run the actual unit test method.
-                method(*args, **kwargs)
-        finally:
-            # Re-run engine discovery as the settings have changed.
-            manager.discover_engines()
 
-    return wrapper
+@wrapt.decorator
+def with_null_executor(wrapped_method, instance, args, kwargs):
+    """Decorate unit test to run processes with the Null executor."""
+    return with_custom_executor(NAME='resolwe.flow.executors.null')(wrapped_method)(*args, **kwargs)
+
+
+@wrapt.decorator
+def with_resolwe_host(wrapped_method, instance, args, kwargs):
+    """Decorate unit test to give it access to a live Resolwe host.
+
+    Set ``RESOLWE_HOST_URL`` setting to the address where the testing
+    live Resolwe host listens to.
+
+    .. note::
+
+        This decorator must be used with a (sub)class of
+        :class:`~django.test.LiveServerTestCase` which starts a live
+        Django server in the background.
+
+    """
+    if not hasattr(instance, 'server_thread'):
+        raise AttributeError(
+            "with_resolwe_host decorator must be used with a "
+            "(sub)class of LiveServerTestCase that has the "
+            "'server_thread' attribute"
+        )
+    resolwe_host_url = 'http://{host}:{port}'.format(
+        host=instance.server_thread.host, port=instance.server_thread.port)
+    with override_settings(RESOLWE_HOST_URL=resolwe_host_url):
+        # Run the actual unit test method.
+        return with_custom_executor(NETWORK='host')(wrapped_method)(*args, **kwargs)